Creates my Octopus (ViemModel/Controller)
This helps with communication between my model and my intended view.

createStudent helps me create an object of student that contains the student name, array of attendance and, number of attendance. It uses the number of attendance days (in the model) to generate false and input it in the attendance array.

addStudent function helps create and add  student to the model studentList array.

getNumOfAttendDays function returns the number of attendance in the model.

setNumOfAttendDays function - this set the number of attendance in the model.

getSelectedStudent function - this returns selected student.

setSelectedStudent - this sets student from the fed varible.

getStudentList - gets the studentList from the model

updateSelectedStudentAttendance - updates selected student atendance record.

updateSelectedName - updates selected student name